requirements.txt
fastapi==0.115.6
flet==0.26.0
mysql_connector_repackaged==0.3.1
pydantic==2.10.5
PyJWT==2.6.0
python_dateutil==2.9.0.post0
qrcode==7.4.2
Requests==2.32.3
pymysql==1.1.1
SQLAlchemy==1.3.20
uvicorn==0.34.0
Werkzeug==1.0.1